added some minor fixes from reserve_lock; fixed all warnings;
Former-commit-id: eb4f457157340b23179077816a8bd6f3d4366fe5

+++ trunk/src/coding/cauchycoding.cc
@@ -38,7 +38,7 @@ vector<BlockData> CauchyCoding::encode(SegmentData segmentData, string setting) 
 
 
 	if (k <= 0 || m < 0 || w <= 0 || w > 32
-			|| (w < 30 && k + m > (1 << w))) {
+			|| (w < 30 && k + m > ((uint32_t) 1 << w))) {
 		cerr << "Bad Parameters" << endl;
 		exit(-1);
 	}
